requirements.txt
boto3==1.12.13
botocore==1.15.13
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.6
django-heroku==0.3.1
django-paypal==1.0.0
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
jmespath==0.9.5
Pillow==7.0.0
psycopg2==2.8.4
pycryptodome==3.9.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
stripe==2.43.0
urllib3==1.25.8
whitenoise==5.0.1
